+++ src/Converter/DBAL/LikeQueryExprConverter.php
@@ -1,6 +1,7 @@
 <?php
-namespace Oka\PaginationBundle\Converter;
+namespace Oka\PaginationBundle\Converter\DBAL;
 
+use Oka\PaginationBundle\Converter\AbstractQueryExprConverter;
 use Oka\PaginationBundle\Exception\BadQueryExprException;
 
 /**
@@ -8,9 +9,9 @@ use Oka\PaginationBundle\Exception\BadQueryExprException;
  * @author Cedrick Oka Baidai <okacedrick@gmail.com>
  *
  */
-class NotLikeQueryExprConverter extends AbstractQueryExprConverter
+class LikeQueryExprConverter extends AbstractQueryExprConverter
 {
-	const PATTERN = '#^notlike\((.+)\)$#i';
+	const PATTERN = '#^like\((.+)\)$#i';
 	
 	/**
 	 * {@inheritdoc}
@@ -22,14 +23,14 @@ class NotLikeQueryExprConverter extends AbstractQueryExprConverter
 		preg_match(self::PATTERN, $exprValue, $matches);
 		
 		if (!preg_match(self::PATTERN, $exprValue, $matches)) {
-			throw new BadQueryExprException(sprintf('the not like query expression converter does not support the following pattern "%s".', $exprValue));
+			throw new BadQueryExprException(sprintf('The like query expression converter does not support the following pattern "%s".', $exprValue));
 		}
 		
 		$value = $matches[1];
 		
 		return $dbDriver === 'orm' ? 
-				(new \Doctrine\ORM\Query\Expr())->notLike($alias.'.'.$field, $namedParameter ?: ':'.$field) : 
-				(new \Doctrine\MongoDB\Query\Expr())->not((new \Doctrine\MongoDB\Query\Expr())->field($field)->text($value));
+				(new \Doctrine\ORM\Query\Expr())->like($alias.'.'.$field, $namedParameter ?: ':'.$field) : 
+				(new \Doctrine\MongoDB\Query\Expr())->field($field)->text($value);
 	}
 	
 	/**